pm2 upgrades

+++ cogs/weed.py
@@ -11,16 +11,19 @@ from datetime import datetime
 class weed(commands.Cog):
     def __init__(self, client):
         self.client = client
-        self.blink = Blink()
-        self.auth = Auth({"username": BLINK_USERNAME,
-                         "password": BLINK_PASSWORD}, no_prompt=True)
-        self.blink.auth = self.auth
-        self.blink.start()
+        if PY_ENV == 'PROD':
+            self.blink = Blink()
+            self.auth = Auth({"username": BLINK_USERNAME,
+                            "password": BLINK_PASSWORD}, no_prompt=True)
+            self.blink.auth = self.auth
+            self.blink.start()
 
     @commands.Cog.listener()
     async def on_ready(self):
-        pass
-        self.weed_loop.start()
+        if PY_ENV == 'PROD':
+            self.weed_loop.start()
+        else:
+            pass
 
     async def weed_helper(self, msg):
         camera = self.blink.cameras['Wiz']
@@ -29,33 +32,33 @@ class weed(commands.Cog):
         self.blink.refresh()
         current_time = datetime.now().timestamp()
         image_name = f'grow-log-{current_time}'
-        camera.image_to_file(f'./img/grow_log/{image_name}.jpg')
+        camera.image_to_file(
+            f'{os.getcwd()}/jeff-bezos-bot/img/grow_log/{image_name}.jpg')
         embed = Embed(
             title=msg,
             color=Color.green())
-        file = File(f'./img/grow_log/{image_name}.jpg', filename='image.jpg')
+        file = File(
+            f'{os.getcwd()}/jeff-bezos-bot/img/grow_log/{image_name}.jpg', filename='image.jpg')
         embed.set_image(url='attachment://image.jpg')
         return file, embed, image_name
 
     @tasks.loop(minutes=60)
     async def weed_loop(self):
-        # await ctx.send("Hourly grow tent updates, sponsored by Kyle Richter")
         msg = "Hourly Grow Tent Update... Sponsored by Daddy Bezos"
         file, embed, _ = await weed.weed_helper(self, msg)
         await self.client.get_channel(LOG_CH_ID).send(file=file, embed=embed)
 
     @commands.command()
     async def weed(self, ctx):
+        if ctx.channel.id == KYLE_LOG_CH_ID:
+            return
         temp = await ctx.send("Please wait..")
         msg = "Here's your pic, you filthy animal."
         file, embed, image_name = await weed.weed_helper(self, msg)
         await ctx.send(file=file, embed=embed)
-        os.remove(f'./img/grow_log/{image_name}.jpg')
+        os.remove(f'{os.getcwd()}/jeff-bezos-bot/img/grow_log/{image_name}.jpg')
         await temp.delete()
 
 
 def setup(client):
     client.add_cog(weed(client))
-
-# for name,camera in blink.cameras.items():
-#     print(name)
